+++ servers/seriesHTTPhandler.go
@@ -2,6 +2,7 @@ package servers
 
 import (
 	"encoding/json"
+	"fmt"
 	"gateserver/storage"
 	"gateserver/support"
 	"log"
@@ -85,12 +86,15 @@ func seriesHTTPhandler() http.Handler {
 				var st, en int64
 				var e error
 				if st, e = strconv.ParseInt(params["start"], 10, 64); e != nil {
+					_, _ = fmt.Fprintf(w, "")
 					return
 				}
 				if en, e = strconv.ParseInt(params["end"], 10, 64); e != nil {
+					_, _ = fmt.Fprintf(w, "")
 					return
 				}
 				if st >= en {
+					_, _ = fmt.Fprintf(w, "")
 					return
 				}
 				var s0, s1 storage.SampleData
@@ -113,7 +117,10 @@ func seriesHTTPhandler() http.Handler {
 				if tag, ts, vals, e := storage.ReadSeriesTS(s0, s1, params["analysis"] != "current"); e == nil {
 					rt = s0.UnmarshalSliceSS(tag, ts, vals)
 				}
-				_ = json.NewEncoder(w).Encode(rt)
+				// the if is added to deal with timeout issues due to the fact this read can eb too long
+				if e = json.NewEncoder(w).Encode(rt); e != nil {
+					_, _ = fmt.Fprintf(w, "")
+				}
 			}
 		}
 	})